LayoutTests/fast/forms/input-stepup-stepdown.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../../fast/js/resources/js-test-pre.js"></script>
</head>
<body>
<p id="description"></p>
<div id="console"></div>
<script>
description('Check stepUp() and stepDown() bahevior for unsupported types.');

var input = document.createElement('input');
var invalidStateErr = '"Error: InvalidStateError: DOM Exception 11"';

debug('Unsupported type');
input.type = 'text';
shouldThrow('input.step = "3"; input.min = ""; input.max = ""; input.value = "2"; input.stepDown()', '"Error: InvalidStateError: DOM Exception 11"');
shouldThrow('input.stepDown(0)', '"Error: InvalidStateError: DOM Exception 11"');
shouldThrow('input.stepUp()', '"Error: InvalidStateError: DOM Exception 11"');
shouldThrow('input.stepUp(0)', '"Error: InvalidStateError: DOM Exception 11"');
</script>
<script src="../../fast/js/resources/js-test-post.js"></script>
</body>
</html>
